+++ src/Responsible.Tests/SelectFromInstructionTests.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Threading.Tasks;
 using NUnit.Framework;
 using Responsible.Tests.Utilities;
 using static Responsible.Responsibly;
@@ -18,35 +19,35 @@ namespace Responsible.Tests
 		}
 
 		[Test]
-		public void SelectFromInstruction_PublishesCorrectError_WhenExceptionThrown()
+		public async Task SelectFromInstruction_PublishesCorrectError_WhenExceptionThrown()
 		{
 			var task = Return(2)
 				.Select<int, int>(_ => throw new Exception("Fail!"))
 				.ToTask(this.Executor);
-			Assert.IsNotNull(GetFailureException(task));
+			Assert.IsNotNull(await AwaitFailureExceptionForUnity(task));
 		}
 
 		[Test]
-		public void SelectFromInstruction_ContainsFailureDetails_WhenFailed()
+		public async Task SelectFromInstruction_ContainsFailureDetails_WhenFailed()
 		{
 			var task = Return(2)
 				.Select<int, int>(_ => throw new Exception("Fail!"))
 				.ToTask(this.Executor);
 
-			var exception = GetFailureException(task);
+			var exception = await AwaitFailureExceptionForUnity(task);
 			StateAssert.StringContainsInOrder(exception.Message)
 				.Failed("SELECT")
 				.FailureDetails();
 		}
 
 		[Test]
-		public void SelectFromInstruction_ContainsCorrectDetails_WhenInstructionFailed()
+		public async Task SelectFromInstruction_ContainsCorrectDetails_WhenInstructionFailed()
 		{
 			var task = DoAndReturn<int>("Throw", () => throw new Exception("Fail!"))
 				.Select(i => i)
 				.ToTask(this.Executor);
 
-			var exception = GetFailureException(task);
+			var exception = await AwaitFailureExceptionForUnity(task);
 			StateAssert.StringContainsInOrder(exception.Message)
 				.Failed("Throw")
 				.FailureDetails()