+++ src/Responsible.Tests/TestInstructionExecutorTests.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using Responsible.Tests.Utilities;
 using static Responsible.Responsibly;
 
 namespace Responsible.Tests
@@ -17,12 +16,11 @@ namespace Responsible.Tests
 		};
 
 		[Test]
-		[TaskExceptionTest]
-		public void Dispose_CancelsInstructions()
+		public async Task Dispose_CancelsInstructions()
 		{
 			var task = Never.ExpectWithinSeconds(1).ToTask(this.Executor);
 			this.Executor.Dispose();
-			var exception = GetFailureException(task);
+			var exception = await AwaitFailureExceptionForUnity(task);
 			Assert.IsInstanceOf<TaskCanceledException>(exception.InnerException);
 		}
 
@@ -35,14 +33,14 @@ namespace Responsible.Tests
 		}
 
 		[Test]
-		public void StartingInstructionTwice_Throws()
+		public async Task StartingInstructionTwice_Throws()
 		{
 			var state = Return(0).CreateState();
 			var task1 = state.ToTask(this.Executor);
 			var task2 = state.ToTask(this.Executor);
 
 			Assert.IsFalse(task1.IsFaulted);
-			var exception = GetFailureException(task2);
+			var exception = await AwaitFailureExceptionForUnity(task2);
 			Assert.IsInstanceOf<InvalidOperationException>(exception.InnerException);
 		}